move tests back to info_cmd.rs and wallet balance fixes
